太阳石与月亮石的暴击时间

+++ src/main/java/org/confluence/mod/item/curio/combat/DestroyerEmblem.java
@@ -23,8 +23,8 @@ public class DestroyerEmblem extends BaseCurioItem implements ICriticalHit {
     }
 
     @Override
-    public float getChance() {
-        return 0.08F;
+    public double getChance() {
+        return 0.08;
     }
 
     @Override
